Upgraded to RBDoom3BFG

+++ doomclassic/doom/d_net.cpp
@@ -40,8 +40,8 @@ If you have questions concerning this license or the applicable additional terms
 
 #include "doomlib.h"
 #include "Main.h"
-#include "../../neo/sys/sys_session.h"
-#include "../../neo/framework/Common_dialog.h"
+#include "d3xp/Game_local.h"
+
 
 void I_GetEvents( controller_t * );
 void D_ProcessEvents (void); 
@@ -60,14 +60,21 @@ extern bool globalNetworking;
 // a ::g->gametic cannot be run until ::g->nettics[] > ::g->gametic for all ::g->players
 //
 
+
+
+
 #define NET_TIMEOUT	1 * TICRATE
+
+
+
+
 
 //
 //
 //
 int NetbufferSize (void)
 {
-	int size = (int)&(((doomdata_t *)0)->cmds[::g->netbuffer->numtics]);
+	int size = (intptr_t)&(((doomdata_t *)0)->cmds[::g->netbuffer->numtics]);
 
 	return size;
 }
@@ -83,7 +90,7 @@ unsigned NetbufferChecksum (void)
 	c = 0x1234567;
 
 	if ( globalNetworking ) {
-		l = (NetbufferSize () - (int)&(((doomdata_t *)0)->retransmitfrom))/4;
+		l = (NetbufferSize () - (intptr_t)&(((doomdata_t *)0)->retransmitfrom))/4;
 		for (i=0 ; i<l ; i++)
 			c += ((unsigned *)&::g->netbuffer->retransmitfrom)[i] * (i+1);
 	}
